stac-gen/requirements.txt
python-cmr
awslambdaric